pyproject.toml
[project]
name = "print-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
